Actualización servicios y repositorios de Taller y Mecánico

+++ src/main/java/com/example/entities/Taller.java
@@ -26,6 +26,6 @@ public class Taller {
     @OneToMany
     @JoinColumn(name = "mecanico_id")
     private List<Mecanico> mecanico;
-    private Double telefono;
+    private Integer telefono;
 
 }
